# Reduce allocations while enumerating Speller suggestions and improve performance