--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Documents/Speller.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Documents/Speller.cs
@@ -177,16 +177,15 @@ namespace System.Windows.Documents
         // for an error range.
         // This method actually runs the speller on the specified text,
         // re-evaluating the error from scratch.
-        internal IList GetSuggestionsForError(SpellingError error)
+        internal List<string> GetSuggestionsForError(SpellingError error)
         {
             ITextPointer contextStart;
             ITextPointer contextEnd;
             ITextPointer contentStart;
             ITextPointer contentEnd;
             TextMap textMap;
-            ArrayList suggestions;
 
-            suggestions = new ArrayList(1);
+            List<string> suggestions = new();
 
             //
             // IMPORTANT!!
@@ -915,7 +914,7 @@ namespace System.Windows.Documents
             {
                 if (textSegment.SubSegments.Count == 0)
                 {
-                    ArrayList suggestions = (ArrayList)data.Data;
+                    List<string> suggestions = (List<string>)data.Data;
                     if(textSegment.Suggestions.Count > 0)
                     {
                         foreach(string suggestion in textSegment.Suggestions)